+++ src/main/java/com/ath/adminefectivo/entities/MaestroLlavesCostosEntity.java
@@ -1,9 +1,5 @@
 package com.ath.adminefectivo.entities;
 
-import java.math.BigInteger;
-import java.util.Date;
-import java.util.List;
-
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.EntityListeners;
@@ -36,10 +32,10 @@ public class MaestroLlavesCostosEntity {
 	@Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_llave")
-    private Long id;
+    private Integer id;
 
     @Column(name = "id_maestro_llave", nullable = false, unique = true)
-    private BigInteger idMaestroLlave;
+    private Long idMaestroLlave;
 
     @Column(name = "estado", length = 20)
     private String estado;
